fixed bugs with data validation and api summary, and some exception handling, outstanding bug is map-reduce returns fewer rows than the file has, perhaps race cnditions on Reduce

+++ src/Models/Bucket.cs
@@ -2,8 +2,8 @@
 {
     public class Bucket
     {
-        public float bucketMin { get; set; }
-        public float bucketMax { get; set; }
+        public double bucketMin { get; set; }
+        public double bucketMax { get; set; }
         public int Count { get; set; }
     }
 }
